backend/requirements.txt
livekit-agents>=0.12.0
livekit-plugins-openai>=0.12.0
livekit-plugins-elevenlabs>=0.12.0
livekit-plugins-anthropic>=0.12.0
livekit-plugins-hedra>=0.12.0
python-dotenv
Pillow
